create reusable custom validation
This commit introduces validation in a more resusable manner. This
commit should be used as a reference for future validation.

Steps for future:
1. For each column you want to validate, create a function in
validators.py called <class-name><variable-name> in camel case
2. Also, if necessary, create a new exception for the case you're doing
3. Perform validation in the function and if it fails raise the
exception that corresponds

Note: this commit also adds a test for the validation, checking that the
expected error is thrown when you enter an invalid value.