Document milestone (#837)
Added a milestone extension file. It's a bit overkill as it's an extension for one field. If we're not adding any more fields to it. We should fold it into the Document model, instead.
Changelog:
Migration to add Document Milestone
Migration to Add the record number to Document
Updated UI so it shows only the fields related to the attachment type